chore: import slicer from remote
