+++ src/containers/Slicer/hooks/useHierarchyTable.ts
@@ -1,6 +1,6 @@
 // create table data for the hierarchy
 import { useGetFolderListQuery } from '@queries/getHierarchy'
-import { TableRow } from '../SlicerTable'
+import { TableRow } from '../types'
 import { useCallback, useMemo } from 'react'
 import { FolderListItem } from '@api/rest/folders'
 import { FolderType } from '@api/rest/project'
